removed key table commands from scala
