+++ python/pyhail/dataset.py
@@ -1,4 +1,5 @@
 from pyhail.java import scala_package_object
+from pyhail.keytable import KeyTable
 
 import pyspark
 
@@ -6,6 +7,18 @@ class VariantDataset(object):
     def __init__(self, hc, jvds):
         self.hc = hc
         self.jvds = jvds
+
+    def aggregate_by_key(self, key_condition = None, agg_condition = None):
+        """Aggregate by user-defined key and aggregation expressions
+
+        :param str key_condition: Named expression for how key
+
+        :param str agg_condition: Named aggregation expression.
+
+        :return: KeyTable
+
+        """
+        return KeyTable(self.hc, self.jvds.aggregateByKey(key_cond, agg_condition))
 
     def aggregate_intervals(self, input, condition, output):
         """Aggregate over intervals and export.
