--- src/main/scala/org/broadinstitute/hail/driver/AnnotateKeyTable.scala
@@ -1,9 +1,0 @@
-package org.broadinstitute.hail.driver
-
-object AnnotateKeyTable extends SuperCommand {
-  def name = "annotatekeytable"
-
-  def description = "Annotate key tables"
-
-  register(AnnotateKeyTableExpr)
-}
