--- src/main/scala/org/broadinstitute/hail/driver/FilterKeyTable.scala
@@ -1,9 +1,0 @@
-package org.broadinstitute.hail.driver
-
-object FilterKeyTable extends SuperCommand {
-  def name = "filterkeytable"
-
-  def description = "Filter key tables"
-
-  register(FilterKeyTableExpr)
-}
